+++ src/client.ts
@@ -25,7 +25,7 @@ import {
   createOAuth2HttpClient,
 } from './http';
 import { validateConfig } from './utils';
-import { OAuthClient } from './http/types';
+import { OAuthClient, TokenDataOauth1 } from './http/types';
 
 export interface OAuth1Authenticate {
   /**
@@ -46,7 +46,7 @@ export interface OAuth1Authenticate {
     token: string;
     tokenSecret: string;
     skipTokenCheck?: boolean;
-  }): Promise<void>;
+  }): Promise<TokenDataOauth1>;
   /**
    * Use OAuth1 Password authentication
    * @example
@@ -61,7 +61,10 @@ export interface OAuth1Authenticate {
    * @throws {TooManyFailedAttemptsError}
    * @throws {MfaRequiredError}
    */
-  authenticate(oauth: { email: string; password: string }): Promise<void>;
+  authenticate(oauth: {
+    email: string;
+    password: string;
+  }): Promise<TokenDataOauth1>;
 }
 
 export interface OAuth2Authenticate {
@@ -182,36 +185,9 @@ export interface Client<T extends ClientParams> {
    * Provides authentication functionality. The Authentication service supports both OAuth 1.0a and OAuth 2.0 standards.
    * @see https://developers.extrahorizon.io/services/?service=auth-service&redirectToVersion=2
    */
-  auth: ReturnType<typeof authService> & {
-    /**
-     *  Confirm MFA method with token, methodId and code
-     *  @example
-     *  try {
-     *    await sdk.auth.authenticate({
-     *      password: '',
-     *      username: '',
-     *    });
-     *  } catch (error) {
-     *    if (error instanceof MfaRequiredError) {
-     *      const { mfa } = error.response;
-     *
-     *      // Your logic to request which method the user want to use in case of multiple methods
-     *      const methodId = mfa.methods[0].id;
-     *
-     *      await sdk.auth.confirmMfa({
-     *        token: mfa.token,
-     *        methodId,
-     *        code: '', // code from ie. Google Authenticator
-     *      });
-     *    }
-     *  }
-     */
-    confirmMfa: (mfa: {
-      token: string;
-      methodId: string;
-      code: string;
-    }) => Promise<void>;
-  } & Authenticate<T>;
+  auth: ReturnType<typeof authService> &
+    Pick<OAuthClient, 'confirmMfa' | 'logout'> &
+    Authenticate<T>;
 }
 
 /**
@@ -252,9 +228,10 @@ export function createClient<T extends ClientParams>(rawConfig: T): Client<T> {
     events: eventsService(httpWithAuth),
     auth: {
       ...authService(httpWithAuth),
-      authenticate: (oauth: AuthParams): Promise<void> =>
+      authenticate: (oauth: AuthParams) =>
         httpWithAuth.authenticate(parseAuthParams(oauth)),
       confirmMfa: httpWithAuth.confirmMfa,
+      logout: httpWithAuth.logout,
     } as any,
     raw: httpWithAuth,
   };